+++ work/services/install_services.py
@@ -1,16 +1,38 @@
-from django.shortcuts import render, redirect
+import random
+import string
+
+from django.utils import timezone
+from django.http import JsonResponse
+from django.shortcuts import get_object_or_404, render, redirect
 from system_manager.models import ConstructManager, InstallLocate
 from work.forms.install_form import InstallCheckListForm
 from django.contrib import messages
 
 from work.models import (
     InspectionItem,
+    InspectionItemCategory,
     InspectionResult,
     InstallCheckList,
-    Measure,
+    MeasureImg,
 )
 
-from ..services.common_services import image_send, sms_send
+from ..services.common_services import sms_send
+
+
+def create_item(type, title, category):
+    if type == "강관 비계":
+        equipment = "1"
+    elif type == "시스템 비계":
+        equipment = "2"
+    else:
+        equipment = "3"
+    new_item = InspectionItem(
+        equipment=equipment,
+        title=title,
+        categoryId=InspectionItemCategory.objects.get(type=category),
+    )
+    new_item.save()
+    return new_item.pk
 
 
 def install_checklist_service(request, type: str):
@@ -31,7 +53,6 @@ def install_checklist_service(request, type: str):
                 pk=request.POST["locate"]
             )
             pk_checklist = request.POST.getlist("checklist-pk")
-            image_keys = request.FILES.keys()
             doing_checklist.save()
             for pk in pk_checklist:
                 result_item = InspectionResult(
@@ -39,12 +60,14 @@ def install_checklist_service(request, type: str):
                     install_checklist_id=doing_checklist,
                     inspection_item_id=InspectionItem.objects.get(pk=pk),
                 )
-                result_item.content = request.POST[f"{pk}-belong"]
                 result_item.save()
-                if f"{pk}-images[]" in image_keys:
+                if request.POST[pk] == "2":
+                    result_measure = result_item.measure.create(
+                        content=request.POST[f"{pk}-belong"]
+                    )
                     images = request.FILES.getlist(f"{pk}-images[]")
                     for img in images:
-                        result_item.measures.create(img=img)
+                        result_measure.measure_imgs.create(img=img)
             messages.success(request, "저장이 완료되었습니다.")
             return redirect(
                 "work:update_install_checklist",
@@ -53,9 +76,10 @@ def install_checklist_service(request, type: str):
             )
     else:
         form = InstallCheckListForm()
-    checklist = InspectionItem.objects.filter(equipment=equipment).order_by(
-        "categoryId__pk"
-    )
+    checklist = InspectionItem.objects.filter(
+        equipment=equipment,
+        init_item=True,
+    ).order_by("categoryId__pk")
     last_doc = InstallCheckList.objects.last()
     if not last_doc:
         docNum = 1
@@ -78,6 +102,7 @@ def update_checklist_service(request, type, pk):
     if request.method == "POST":
         form = InstallCheckListForm(request.POST, instance=instance)
         if form.is_valid():
+            # 기본 저장 영역
             doing_checklist = form.save(commit=False)
             doing_checklist.equipment = type
             if "locate" in request.POST.keys():
@@ -85,35 +110,52 @@ def update_checklist_service(request, type, pk):
                     pk=request.POST["locate"]
                 )
             pk_checklist = request.POST.getlist("checklist-pk")
-            image_keys = request.FILES.keys()
             doing_checklist.save()
+
+            # 삭제 영역
+            install_list = doing_checklist.inspection_result.all()
+            for install in install_list:
+                if str(install.inspection_item_id.pk) not in pk_checklist:
+                    install.delete()
+
+            # 수정 영역
             for pk_item in pk_checklist:
                 before_result_item = InspectionResult.objects.filter(
                     install_checklist_id=doing_checklist,
-                    inspection_item_id=InspectionItem.objects.get(pk=pk_item),
-                )[0]
+                    inspection_item_id=InspectionItem.objects.get(
+                        pk=pk_item,
+                    ),
+                )
+                delete_before = False
+                # 수정된 영역이면 해당 구문이 실행, 새로 생성된 영역은 패스
+                if before_result_item:
+                    before_result_item = before_result_item[0]
+                    delete_before = True
                 result_item = InspectionResult(
                     result=request.POST[pk_item],
                     install_checklist_id=doing_checklist,
-                    content=request.POST[f"{pk_item}-belong"],
                     inspection_item_id=InspectionItem.objects.get(pk=pk_item),
                 )
                 result_item.save()
-                image_ids = request.POST.getlist(f"{pk_item}-images-preloaded[]")
-                for image_id in image_ids:
-                    image = Measure.objects.filter(
-                        pk=image_id,
-                        inspectionResult=before_result_item,
+                if request.POST[pk_item] == "2":
+                    result_measure = result_item.measure.create(
+                        content=request.POST[f"{pk_item}-belong"]
                     )
-                    if image:
-                        image = image[0]
-                        image.inspectionResult = result_item
-                        image.save()
-                if f"{pk_item}-images[]" in image_keys:
+                    before_image_ids = request.POST.getlist(
+                        f"{pk_item}-images-preloaded[]"
+                    )
+                    for before_image_id in before_image_ids:
+                        before_image = MeasureImg.objects.get(
+                            pk=before_image_id,
+                        )
+                        if before_image:
+                            before_image.measure = result_measure
+                            before_image.save()
                     images = request.FILES.getlist(f"{pk_item}-images[]")
                     for img in images:
-                        result_item.measures.create(img=img)
-                before_result_item.delete()
+                        result_measure.measure_imgs.create(img=img)
+                if delete_before:
+                    before_result_item.delete()
             messages.success(request, "저장이 완료되었습니다.")
             return redirect("work:update_install_checklist", type, pk)
     else:
@@ -131,25 +173,137 @@ def update_checklist_service(request, type, pk):
 
 
 def assign_cm(request, type):
-    cm = ConstructManager.objects.get(pk=request.POST.get("sign"))
     doc = InstallCheckList.objects.get(pk=request.POST.get("docNum"))
-    doc.isSuccess = True
+    doc.isCheckWriter = True
+    doc.isCheckCM = False
+    if not doc.cm:
+        cm = ConstructManager.objects.get(pk=request.POST.get("sign"))
+        doc.cm = cm
+    else:
+        cm = doc.cm
+    doc.expired_date = request.POST.get("expired_date")
+    doc.urlCode = "".join(
+        random.choices(
+            string.ascii_letters + string.digits,
+            k=12,
+        )
+    )
     doc.save()
     link = request.build_absolute_uri(f"/work/read_install/{type}/{doc.pk}/")
     cm_phone = cm.phone
     sms_send(link, [cm_phone], 3)
-    checklist_ids = request.POST.getlist("checklist_id")
-    message_list = []
-    for checklist_id in checklist_ids:
-        target = InspectionResult.objects.get(
-            install_checklist_id=doc,
-            inspection_item_id=InspectionItem.objects.get(pk=checklist_id),
-        )
-        message_list.append(
-            {
-                "content": target.content,
-                "img": list(map(lambda x: x.img, list(target.measures.all()))),
-            }
-        )
-    image_send(message_list, cm_phone)
+    # checklist_ids = request.POST.getlist("checklist_id")
+    # message_list = []
+    # for checklist_id in checklist_ids:
+    #     target = InspectionResult.objects.get(
+    #         install_checklist_id=doc,
+    #         inspection_item_id=InspectionItem.objects.get(pk=checklist_id),
+    #     )
+    #     message_list.append(
+    #         {
+    #             "content": target.content,
+    #             "img": list(map(lambda x: x.img, list(target.measures.all()))),
+    #         }
+    #     )
+    # image_send(message_list, cm_phone)
     # 문자 전송 페이지 만들기
+
+
+def delete_install_checklists_service(request):
+    if request.method == "POST":
+        install_checklist_list = request.POST.getlist("delete_list[]")
+        for install_checklist in install_checklist_list:
+            install_checklist = InstallCheckList.objects.get(
+                pk=install_checklist,
+            )
+            install_checklist.delete()
+        return JsonResponse({"result": "success"})
+    return JsonResponse({"result": "fail"}, status=400)
+
+
+def measure_install_service(request, urlCode):
+    checklist = get_object_or_404(
+        InstallCheckList,
+        urlCode=urlCode,
+        isCheckWriter=True,
+    )
+    if request.method == "POST":
+        form_key = list(request.POST.keys())
+        form_key.remove("csrfmiddlewaretoken")
+        for key in form_key:
+            pk = key.split("-")[0]
+            inspection_result = InspectionResult.objects.get(pk=pk)
+            measure = inspection_result.measure.create(
+                isCM=True,
+                content=request.POST[key],
+                cm=checklist.cm,
+            )
+            for image in request.FILES.getlist(pk + "-images[]"):
+                measure.measure_imgs.create(img=image)
+        checklist.isCheckCM = True
+        checklist.isCheckWriter = False
+        checklist.save()
+        return redirect("main:home")
+    if checklist.expired_date < timezone.now():
+        return redirect("main:home")
+    return render(
+        request,
+        "work/install/doing/measure_checklist.html",
+        {
+            "checklist": checklist,
+            "checklist_items": checklist.inspection_result.all(),
+        },
+    )
+
+
+def review_install_checklist_service(request, type, pk):
+    checklist = InstallCheckList.objects.get(pk=pk)
+    isSave = False
+    inspectionResults = checklist.inspection_result.filter(result="2")
+    for inspectionResult in inspectionResults:
+        if not inspectionResult.measure.last().isCM:
+            isSave = True
+            break
+    if request.method == "POST":
+        for inspectionResult in inspectionResults:
+            lastMeasure = inspectionResult.measure.last()
+            if lastMeasure.isCM:
+                newMeasure = inspectionResult.measure.create(
+                    content=request.POST[f"{inspectionResult.pk}-content"],
+                )
+                images = request.FILES.getlist(f"{inspectionResult.pk}-images[]")
+                for img in images:
+                    newMeasure.measure_imgs.create(img=img)
+            else:
+                lastMeasure.content = request.POST[f"{inspectionResult.pk}-content"]
+                lastMeasure.save()
+                before_images = list(
+                    request.POST.getlist(f"{inspectionResult.pk}-images-preloaded[]")
+                )
+                for measure_img in lastMeasure.measure_imgs.all():
+                    if str(measure_img.pk) not in before_images:
+                        measure_img.delete()
+                images = request.FILES.getlist(f"{inspectionResult.pk}-images[]")
+                for img in images:
+                    lastMeasure.measure_imgs.create(img=img)
+        messages.success(request, "저장이 완료되었습니다.")
+        isSave = True
+    return render(
+        request,
+        "work/install/doing/review_install_checklist.html",
+        {
+            "type": type,
+            "checklist": checklist,
+            "checklist_items": checklist.inspection_result.all(),
+            "isSave": isSave,
+        },
+    )
+
+
+def success_install_checklist_service(pk):
+    checklist = get_object_or_404(InstallCheckList, pk=pk)
+    checklist.isSuccess = True
+    checklist.isCheckCM = True
+    checklist.isCheckWriter = True
+    checklist.save()
+    return JsonResponse({"result": "success"})
